chore: fix code structure to make test run
